+++ paper-examples/solve_synthetic_with_dcqp_eliminate.m
@@ -1,4 +1,4 @@
-function solve_synthetic_with_dcqp(group_name,instance_id)
+function solve_synthetic_with_dcqp_eliminate(group_name,instance_id)
 
 
 valid_groups = {'qp_n_0_1', 'qp_n_0_3', 'qp_n_0_9', 'qp_u_0_1', 'qp_u_0_3', 'qp_u_0_9', 'qp_u_25_1'};
@@ -9,14 +9,14 @@ end
 
 
 
-diary diaryfile-synthetic-dcqp.txt
+diary diaryfile-synthetic-dcqp-eliminate.txt
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 
 
 datalocation='../data/synthetic/';
-myrecord=zeros(20,7);
+myrecord=zeros(20,6);
 
 fprintf("**************************************start to solve ")
 fprintf(group_name);
@@ -39,6 +39,7 @@ for i=1:length(K)
     filename=[group_name '_' num2str(k)];
 
     datafilenamek=[datalocation filename '.mat'];
+    filename=[filename '_eliminate'];
     F=load(datafilenamek);
 
     H=F.H;
@@ -68,30 +69,46 @@ for i=1:length(K)
     params.nb_rounds=10;
     params.filename=filename;
     params.do_scaling=true;
-    params.tol_mosek = 1e-9; 
     params.verbose=true;
+    %params.eta=0.1;
+    %params.mosek_tolerance = 1e-10;
+
+    fprintf('eta=%.2e\n',params.eta);
+    
+    [Q_tilde, d_tilde, A_tilde, b_tilde, M, x0,obj_constant] = eliminate_equalities(Q, d, A_bar, b_bar, Aeq, beq,n);
+
+
+    fprintf('obj_constant=%.5f\n',obj_constant);
+
+
+    [x_opt, fval, info] = dcqp_solve(Q_tilde, d_tilde, A_tilde, b_tilde, [], [], params);
+    
     
 
-    [x_opt, fval, info] = dcqp_solve(Q, d, A_bar, b_bar, Aeq, beq, params);
+    %[x_opt, fval, info] = dcqp_solve(Q, d, A_bar, b_bar, Aeq, beq, params);
      
     if info.status=="not solved"
        params.eta=0.8;
-       [x_opt, fval, info] = dcqp_solve(Q, d, A_bar, b_bar, Aeq, beq, params);
-     
+       params.gap_tolerance = 1e-5;  
+       [x_opt, fval, info] = dcqp_solve(Q_tilde, d_tilde, A_tilde, b_tilde, [], [], params);
+       
     end
+
+    x_real=M * x_opt + x0;
+    f_real=obj_constant+fval;
+    fprintf('optimal value=%.6f\n', f_real);
     myrecord(k,1)=info.gap;
-    myrecord(k,2)=max(A_bar*x_opt-b_bar);
+    myrecord(k,2)=max(A_bar*x_real-b_bar);
     if ~isempty(beq)
-        myrecord(k,3)=norm(Aeq*x_opt-beq);
+        myrecord(k,3)=norm(Aeq*x_real-beq);
     end
-    myrecord(k,4)=fval;
+    myrecord(k,4)=f_real;
     myrecord(k,5)=info.lower_bound;
     myrecord(k,6)=info.time;
-    myrecord(k,7)=info.iterations;
 end
 
 if length(K)==20 
-    save(['summary_results/' group_name '.mat'],"myrecord");
+    save(['summary_results/' group_name '_eliminate.mat'],"myrecord");
 end
 
 end